requirements.txt
Django==1.11.29
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
Django==1.11.29
django-tinymce==3.0.2
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
